backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# AI/ML - FREE TIER OPTIMIZED
groq==0.4.2
httpx==0.26.0

# Medical NLP
spacy==3.7.2
scispacy==0.5.3

# Database
supabase==2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Vector Store
pinecone-client==3.0.0

# PHI Scrubbing
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Storage
boto3==1.34.34

# Monitoring & Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Code Quality
black==24.1.1
flake8==7.0.0
